Remove unnecessary cast in setScaleFactor overwrite
Eliminated the redundant cast to int when assigning the result of ScaleFactorUtil.fromInternalScaleFactor in the setScaleFactor method.

+++ src/main/java/net/rosemods/betteruiscale/mixin/MixinWindow.java
@@ -41,7 +41,7 @@ public class MixinWindow {
      */
     @Overwrite
     public void setScaleFactor(int scaleFactor) {
-        this.scaleFactor = (int) ScaleFactorUtil.fromInternalScaleFactor(scaleFactor);
+        this.scaleFactor = ScaleFactorUtil.fromInternalScaleFactor(scaleFactor);
         scaledWidth = ScaleFactorUtil.scaleInternal(framebufferWidth, scaleFactor);
         scaledHeight = ScaleFactorUtil.scaleInternal(framebufferHeight, scaleFactor);
     }
